Don't try to wrap when only one return argument
crlfmt would previously try to wrap a line like

    func foo() func(veryLoooooooooooooooooooooooooongReturnType int, anotherArg string)

onto two lines, like so:

    func foo()
        func(veryLoooooooooooooooooooooooooongReturnType int, anotherArg string)

This is a syntax error. Introducing parens, a la

    func foo() (
        func(veryLoooooooooooooooooooooooooongReturnType int, anotherArg string),
    )

would fix the syntax error, but gofmt helpfully "simplifies" that back to the
one line version.

Teach crlfmt to simply not wrap in this case to avoid introducing syntax errors
or fighting with gofmt.