add -srcdir option similar to goimports
The -srcdir option allows goimports to resolve imports as if the file
being processed was located in the given directory. This is useful for
some editor integrations that pass crlfmt a temporary file located in a
different directory than the original source file.

For emacs users, this change gets a bit closer to being able to easily
integrate with go-mode. The final step is either a change to
go-mode.el or a monkey patch to the relevant functions that add the
-srcdir option.

Fixes #28