ignore nonexistant files/directories
The call to filepath.Walk may surface an error if it observes a directory but
attempting to stat it finds that it no longer exists. This type of situation is
possible when running go tests for multiple packages in parallel, and some go
tests create directories within the source tree. This behavior was observed in
the Pebble metamorphic tests.