add flag to wrap function doc strings
This change adds the program flag 'wrapfndoc' which wraps function
doc strings. By default, the doc strings are wrapped to 80 chars.

Noteably, this flag ignores multiline comments and does not wrap
words which are over 80 chars. This is mainly done for simplicity
and can be changed in the future.

Also, it only wraps lines in which are too long and preserves
lines which fit within 80 characters. This is different than
the `Wrap To Column` extension in Goland which greedily writes
80 chars at a time. The problem with writing chars greedily is
that it can mess up formatting. For example, a list of bullet
points gets put onto one line using `Wrap To Column`:

list:
- foo
- bar
- bar

would become

list: - foo - bar - baz

This is undesirable.

It is also worth noting that this change will not change any lines
after `Wrap To Column` is used because it ignores lines that are <=
80 chars.